Decomposition in progress.
