+++ .gitignore
@@ -130,3 +130,7 @@ dmypy.json
 
 # PyCharm
 .idea
+
+# Develop enviroment
+env.py
+./env.py
