+++ step/courses/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
 
-# Register your models here.
+from courses.models import Course
+
+admin.site.register(Course)
