+++ step/courses/templates/course_confirm_delete.html
@@ -16,7 +16,7 @@
   <form method="post">
     {% csrf_token %}
     <button class="btn btn-outline-danger" type="submit">Yes</button>
-    <a class="btn btn-outline-success" href="{% url 'students_list' %}">No</a>
+    <a class="btn btn-outline-success" href="{% url 'employees_list' %}">No</a>
   </form>
 
 {% endblock %}
