+++ step/courses/templates/courses_list.html
@@ -56,7 +56,7 @@
         <button type="button"
                 class="btn btn-outline-warning my-3"
                 onclick="GenerateData()"
-                id="Generate_new_bills">
+                id="GenerateData">
             Generate data
         </button>
         <script>
