+++ step/courses/templates/index.html
@@ -1,0 +1,34 @@
+<!--"""-->
+<!--Шаблон основной страницы.-->
+<!--"""-->
+
+{% extends "base.html" %}
+{% block title %}
+    Main page
+{% endblock %}
+{% block body %}
+    <div class="h2">Greetings, {{ user.username|default:"Guest:)" }}!</div>
+    <div>
+        {% if courses %}
+            <div class="h3">Our courses:</div>
+            <ul class="list-group">
+                {% for course in courses %}
+                    <li class="list-group-item">
+                        <a class="text-decoration-none"
+                           href="{% url 'course_detail' pk=course.pk %}"
+                        >{{ course.name }}</a>
+                        <div>Is it available:
+                            {% if course.is_available %}
+                                <em>Yes</em>
+                            {% else %}
+                                <em>No</em>
+                            {% endif %}
+                        </div>
+                    </li>
+                {% endfor %}
+            </ul>
+        {% else %}
+            <h3>No courses yet</h3>
+        {% endif %}
+    </div>
+{% endblock %}
