+++ step/stafftrain/admin.py
@@ -4,7 +4,8 @@
 
 from django.contrib import admin
 
-from stafftrain.models import Course, Result
+# from stafftrain.models import Course, Result
+from stafftrain.models import Result
 
-admin.site.register(Course)
+# admin.site.register(Course)
 admin.site.register(Result)
