+++ step/stafftrain/management/commands/generate_data.py
@@ -6,12 +6,14 @@ import os
 import random
 
 import requests
-from django.contrib.auth.models import User
 from django.core.exceptions import BadRequest
 from django.core.management.base import BaseCommand
 from django.db import IntegrityError
 
-from stafftrain.models import Course, Result
+import env
+from courses.models import Course
+from employees.models import Employee
+from stafftrain.models import Result
 
 USERS_DATA_URL = "https://jsonplaceholder.typicode.com/users"
 course_list = [
@@ -53,14 +55,14 @@ class Command(BaseCommand):
         print("Erasing previous data")
         Result.objects.all().delete()
         Course.objects.all().delete()
-        User.objects.all().delete()
+        Employee.objects.all().delete()
 
         # 2. Generate base data
         print("Generating base data")
-        super_user = User.objects.create_superuser(
+        super_user = Employee.objects.create_superuser(
             username="macushiro",
             email="macushiro@newbie.com",
-            password=os.environ.get("super_user"),
+            password=env.super_user,
         )
 
         # 2.1 Getting data from external service
@@ -70,34 +72,33 @@ class Command(BaseCommand):
             rand = random
             for elem in data:
                 try:
-                    user = User.objects.create(
+                    employee = Employee.objects.create(
                         username=elem["username"],
                         first_name=elem["name"],
                         email=elem["email"],
                     )
                     # 2.2 Generate courses data
-                    print(user.id, user.username, user.first_name)
+                    print(employee.id, employee.username, employee.first_name)
                     course_name = rand.choice(course_list)
                     course = Course.objects.create(
                         name=course_name,
                         description=f"{course_name} {rand.choice(course_level)} course",
                         is_available=True,
-                        student=user,
                     )
                     print(course.name, course.description)
                 except IntegrityError:
                     raise BadRequest(f"Couldn't delete previous data from Database.")
 
         # 3. Generate results data  -  for feature working
-        # min_user_id = User.objects.order_by('id').first()
-        # max_user_id = User.objects.order_by('id').last()
-        # user_list = range(min_user_id, max_user_id + 1)
+        # min_employee_id = Employee.objects.order_by('id').first()
+        # max_employee_id = Employee.objects.order_by('id').last()
+        # employee_list = range(min_employee_id, max_employee_id + 1)
         # left_bord = 1
-        # for elem in user_list:
+        # for elem in employee_list:
         #     right_bord = left_bord + rand.randrange(1, 10)
         #     while left_bord < right_bord:
         #         result = Result.objects.create(
-        #             student=elem,
+        #             employee=elem,
         #             course=f"Result #{left_bord} for {rand.choice(['normal', 'good', 'great'])}",
         #             percent=rand.uniform(1.0, 100.0),
         #             test_result=left_bord,
