+++ step/stafftrain/models.py
@@ -4,27 +4,27 @@
 
 from django.db import models
 
-# from users.models import Student  - for future functional extending
-from django.contrib.auth.models import User
+from courses.models import Course
+from employees.models import Employee
 
 
-class Course(models.Model):
-    """
-    Course model description
-    """
-    name = models.CharField(
-        verbose_name="Наименование курса/Course name",
-        max_length=256,
-        blank=False,
-        editable=True,
-    )
-    description = models.TextField(
-        verbose_name="О курсе/Description", blank=True, null=True
-    )
-    is_available = models.BooleanField(
-        verbose_name="Доступен/Available", blank=True, null=True
-    )
-    student = models.ForeignKey(User, on_delete=models.PROTECT, null=True)
+# class Course(models.Model):
+#     """
+#     Course model description
+#     """
+#     name = models.CharField(
+#         verbose_name="Наименование курса/Course name",
+#         max_length=256,
+#         blank=False,
+#         editable=True,
+#     )
+#     description = models.TextField(
+#         verbose_name="О курсе/Description", blank=True, null=True
+#     )
+#     is_available = models.BooleanField(
+#         verbose_name="Доступен/Available", blank=True, null=True
+#     )
+#     student = models.ForeignKey(User, on_delete=models.PROTECT, null=True)
 
 
 class Result(models.Model):
@@ -42,5 +42,5 @@ class Result(models.Model):
     )
     # student = models.OneToOneField(Student, on_delete=models.PROTECT)
     # - for future functional extending
-    student = models.OneToOneField(User, on_delete=models.PROTECT)
+    student = models.OneToOneField(Employee, on_delete=models.PROTECT)
     course = models.OneToOneField(Course, on_delete=models.PROTECT)
