+++ step/step/settings.py
@@ -13,6 +13,8 @@ import os
 from pathlib import Path
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
+import env
+
 BASE_DIR = Path(__file__).resolve().parent.parent
 
 
@@ -20,7 +22,8 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # See https://docs.djangoproject.com/en/4.1/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = os.environ.get("SECRET_KEY")
+# SECRET_KEY = os.environ.get("SECRET_KEY")
+SECRET_KEY = env.SECRET_KEY
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
@@ -40,7 +43,8 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     # Core applications
     "stafftrain",
-    "users",
+    "employees",
+    "courses",
 ]
 
 MIDDLEWARE = [
@@ -128,5 +132,6 @@ DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 
 # Authentication
 LOGIN_REDIRECT_URL = "/"
-LOGIN_URL = "/users/login/"
+LOGIN_URL = "/employees/login/"
 LOGOUT_REDIRECT_URL = "/"
+AUTH_USER_MODEL = "employees.Employee"
