+++ step/templates/base.html
@@ -31,11 +31,11 @@
                 <a class="nav-link" aria-current="page" href="{% url 'main_page' %}">Home</a>
                 <a class="nav-link" aria-current="page" href="{% url 'courses_list' %}">Courses</a>
                 {% if user.is_superuser or user.is_staff %}
-                    <a class="nav-link" href="{% url 'students_list' %}">Students</a>
+                    <a class="nav-link" href="{% url 'employees_list' %}">Employees</a>
 {#                    <a class="nav-link" href="{% url 'results' %}">Results</a>#}
                 {% endif %}
                 {% if user.is_authenticated %}
-                    <a class="nav-link" href="{% url 'user_info' %}">User account</a>
+                    <a class="nav-link" href="{% url 'employee_info' %}">User account</a>
                     <a class="nav-link" href="{% url 'logout' %}">Log out</a>
                 {% else %}
                     <a class="nav-link" href="{% url 'login' %}">Log in</a>
