--- step/users/admin.py
@@ -1,9 +1,0 @@
-"""
-    Файл регистрации моделей.
-"""
-
-from django.contrib import admin
-
-# from users.models import Student
-
-# admin.site.register(Student)
