--- step/users/models.py
@@ -1,33 +1,0 @@
-"""
-    Файл моделей приложения.
-"""
-
-from django.db import models
-
-# from django.contrib.auth.models import AbstractUser
-#
-#
-# class Student(AbstractUser):
-#     first_name = models.CharField(
-#         verbose_name="Имя/Name",
-#         max_length=128,
-#         blank=True,
-#         null=True,
-#     )
-#     last_name = models.CharField(
-#         verbose_name="Фамилия/Last name",
-#         max_length=128,
-#         blank=True,
-#         null=True,
-#     )
-#     position = models.CharField(
-#         verbose_name="Должность/Position",
-#         max_length=128,
-#         blank=True,
-#         null=True,
-#     )
-#     about = models.TextField(
-#         verbose_name="О себе/About your self",
-#         blank=True,
-#         null=True,
-#     )
