--- step/users/templates/login_form.html
@@ -1,13 +1,0 @@
-<!--"""-->
-<!--Шаблон для входа пользователя.-->
-<!--"""-->
-
-{% extends 'base.html' %}
-
-{% block body %}
-    <form method="post">
-      {% csrf_token %}
-      {{form.as_p}}
-      <button class="btn btn-outline-success my-3" type="submit">Login</button>
-    </form>
-{% endblock %}
