--- step/users/templates/user_detail.html
@@ -1,25 +1,0 @@
-<!--"""-->
-<!--Шаблон детальной профиля пользователя.-->
-<!--"""-->
-
-{% extends "base.html" %}
-{% block title %}
-    Personal info
-{% endblock %}
-{% block body %}
-    <div class="h3">Personal data</div>
-    <p>
-        <strong>Login as {{ object.username }}</strong>
-    </p>
-    <div>
-        <ul>
-            <li><strong>Почта/Email: </strong>{{ object.email }}</li>
-            <li><strong>Имя/Name: </strong>{{ object.first_name }}</li>
-            <li><strong>Фамилия/Last Name: </strong>{{ object.last_name }}</li>
-        </ul>
-    </div>
-    <a class="btn btn-outline-warning" href="{% url 'user_update' pk=user.pk %}">Update info</a>
-    <a class="btn btn-outline-primary my-3"
-       href="{% url 'courses_list' %}"
-    >Back to courses list</a>
-{% endblock %}
