--- step/users/templates/user_form.html
@@ -1,13 +1,0 @@
-<!--"""-->
-<!--Шаблон/форма регистрации пользователя.-->
-<!--"""-->
-
-{% extends 'base.html' %}
-
-{% block body %}
-    <form method="post">
-      {% csrf_token %}
-      {{form.as_p}}
-      <button class="btn btn-outline-success my-3" type="submit">Registration</button>
-    </form>
-{% endblock %}
