--- step/users/tests.py
@@ -1,7 +1,0 @@
-"""
-    Файл тестирования моделей.
-"""
-
-from django.test import TestCase
-
-# Create your tests here.
